requirements.txt
Django==1.8.4
django-bootstrap3==6.2.2
django-braces==1.8.1
django-crispy-forms==1.5.2
pytz==2015.4
six==1.9.0
wheel==0.24.0
